requirements.txt
opencv-python~=4.1.2.30
pymysql~=1.0.2
numpy
requests~=2.28.1
pillow~=9.2.0
django~=3.0.1
djangorestframework~=3.11.0
djangorestframework-jwt~=1.11.0
setuptools~=65.3.0
django-cors-headers
mysqlclient
aliyun-python-sdk-core
django-redis
drf-extensions
scikit-image